requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
botocore==1.35.60
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.19.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyodbc==5.2.0
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.27
sqlparse==0.4.4
streamlit==1.31.1
streamlit-option-menu==0.3.12
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
vanna==0.1.1
watchdog==4.0.0
Werkzeug==3.0.1
zipp==3.17.0
